+++ kafka-connect-http-connectors/src/main/java/io/github/clescot/kafka/connect/http/client/okhttp/OkHttpClient.java
@@ -7,10 +7,7 @@ import com.google.common.jimfs.Jimfs;
 import io.github.clescot.kafka.connect.http.client.AbstractHttpClient;
 import io.github.clescot.kafka.connect.http.client.Configuration;
 import io.github.clescot.kafka.connect.http.client.HttpException;
-import io.github.clescot.kafka.connect.http.client.okhttp.configuration.AuthenticationConfigurer;
-import io.github.clescot.kafka.connect.http.client.okhttp.configuration.AuthenticationsConfigurer;
-import io.github.clescot.kafka.connect.http.client.okhttp.configuration.BasicAuthenticationConfigurer;
-import io.github.clescot.kafka.connect.http.client.okhttp.configuration.DigestAuthenticationConfigurer;
+import io.github.clescot.kafka.connect.http.client.okhttp.authentication.*;
 import io.github.clescot.kafka.connect.http.client.okhttp.event.AdvancedEventListenerFactory;
 import io.github.clescot.kafka.connect.http.client.okhttp.interceptor.InetAddressInterceptor;
 import io.github.clescot.kafka.connect.http.client.okhttp.interceptor.LoggingInterceptor;
@@ -102,7 +99,8 @@ public class OkHttpClient extends AbstractHttpClient<Request, Response> {
         //authentication
         AuthenticationConfigurer basicAuthenticationConfigurer = new BasicAuthenticationConfigurer();
         AuthenticationConfigurer digestAuthenticationConfigurer = new DigestAuthenticationConfigurer(random);
-        List<AuthenticationConfigurer> authenticatorConfigurers = Lists.newArrayList(basicAuthenticationConfigurer,digestAuthenticationConfigurer);
+        AuthenticationConfigurer oAuth2ClientCredentialsFlowConfigurer = new OAuth2ClientCredentialsFlowConfigurer();
+        List<AuthenticationConfigurer> authenticatorConfigurers = Lists.newArrayList(basicAuthenticationConfigurer,digestAuthenticationConfigurer,oAuth2ClientCredentialsFlowConfigurer);
         AuthenticationsConfigurer authenticationsConfigurer = new AuthenticationsConfigurer(authenticatorConfigurers);
         authenticationsConfigurer.configure(config, httpClientBuilder);
